+++ webgoat-lessons/challenge/src/main/java/org/owasp/webgoat/plugin/ChallengeIntro.java
@@ -1,10 +1,11 @@
 package org.owasp.webgoat.plugin;
 
-import com.google.common.collect.Lists;
+import java.util.List;
+
 import org.owasp.webgoat.lessons.Category;
 import org.owasp.webgoat.lessons.NewLesson;
 
-import java.util.List;
+import com.google.common.collect.Lists;
 
 /**
  * @author nbaars
