+++ webgoat-lessons/sql-injection/src/main/resources/html/SqlInjectionAdvanced.html
@@ -81,7 +81,7 @@
                                 <div class="col-lg-12">
                                     <form id="login-form" class="attack-form" accept-charset="UNKNOWN"
                                           method="POST" name="form"
-                                          action="/WebGoat/SqlInjectionAdvanced/challenge_Login"
+                                          action="/WebGoat/SqlInjectionAdvanced/challenge"
                                           enctype="application/json;charset=UTF-8" role="form">
                                         <div class="form-group">
                                             <input type="text" name="username_login" id="username4" tabindex="1"
